Rename models.[blah].ts to models/[blah].ts
