+++ sdk/eventhub/event-hubs/src/eventHubSender.ts
@@ -25,8 +25,8 @@ import {
 import { EventData, toAmqpMessage } from "./eventData";
 import { ConnectionContext } from "./connectionContext";
 import { LinkEntity } from "./linkEntity";
-import { EventHubProducerOptions } from "./impl/models.private";
-import { SendOptions } from "./models.public";
+import { EventHubProducerOptions } from "./models/private";
+import { SendOptions } from "./models/public";
 
 import { getRetryAttemptTimeoutInMs } from "./util/retries";
 import { AbortSignalLike, AbortError } from "@azure/abort-controller";
