+++ sdk/eventhub/event-hubs/src/models/private.ts
