+++ sdk/eventhub/event-hubs/src/partitionProcessor.ts
@@ -1,5 +1,5 @@
 import { CheckpointStore } from "./eventProcessor";
-import { CloseReason } from "./models.public";
+import { CloseReason } from "./models/public";
 import { ReceivedEventData } from "./eventData";
 import { LastEnqueuedEventProperties } from "./eventHubReceiver";
 import {
